Re-organized package structure to be more logical grouping files based on function
