--- ASWS/src/main/java/com/example/ASWS/ContactRepository.java
@@ -1,7 +1,0 @@
-package com.example.ASWS;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface ContactRepository extends JpaRepository<Contact, Long> {
-
-}
