--- ASWS/src/main/java/com/example/ASWS/CustomerRepository.java
@@ -1,7 +1,0 @@
-package com.example.ASWS;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface CustomerRepository extends JpaRepository<Customer, String> {
-
-}
