+++ ASWS/src/main/java/com/example/ASWS/repositories/CustomerRepository.java
@@ -1,0 +1,8 @@
+package com.example.ASWS.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import com.example.ASWS.models.Customer;
+
+public interface CustomerRepository extends JpaRepository<Customer, String> {
+
+}
